+++ mycodo/databases/models/output.py
@@ -27,6 +27,7 @@ class Output(CRUDMixin, db.Model):
     bit_length = db.Column(db.Integer, default=None)
     on_command = db.Column(db.Text, default=None)
     off_command = db.Column(db.Text, default=None)
+    pwm_command = db.Column(db.Text, default=None)
 
     # PWM
     pwm_hertz = db.Column(db.Integer, default=None)  # PWM Hertz